+++ pom.xml
@@ -16,6 +16,7 @@
     <kotlin.version>1.3.72</kotlin.version>
     <serialization.version>0.20.0</serialization.version>
     <vertx.version>3.9.0</vertx.version>
+    <vertx.jooq.version>5.1.1</vertx.jooq.version>
     <jackson.version>2.10.2</jackson.version>
     <slf4j.version>1.7.25</slf4j.version>
 
@@ -73,7 +74,8 @@
   <modules>
     <module>crabzilla-core</module>
     <module>crabzilla-pg-client</module>
-    <module>crabzilla-web-pg-client</module>
+    <module>crabzilla-pg-client-jooq</module>
+    <module>crabzilla-pg-client-web</module>
     <module>coverage-report</module>
   </modules>
 
@@ -109,6 +111,14 @@
         <version>${vertx.version}</version>
         <type>pom</type>
         <scope>import</scope>
+      </dependency>
+
+      <!-- vertx-jooq -->
+
+      <dependency>
+        <groupId>io.github.jklingsporn</groupId>
+        <artifactId>vertx-jooq-classic-reactive</artifactId>
+        <version>5.1.1</version>
       </dependency>
 
       <!-- util -->
